+++ tests/hierarchy/test_parent_resolution.py
@@ -1,0 +1,131 @@
+import promising
+
+
+async def test_async_context_decorator_resolves_parent_at_call_site() -> None:
+    """
+    The parent of a @promising.context-decorated async function's context
+    is determined at call-site (when the coroutine object is created),
+    not at await-site (when the coroutine body runs).
+
+    Scenario: coroutine created inside `outer`, awaited outside it.
+    func_ctx should still have `outer` as its parent.
+    """
+    func_ctx = None
+
+    @promising.context
+    async def work() -> str:
+        nonlocal func_ctx
+        func_ctx = promising.get_active_context()
+        return "done"
+
+    with promising.context() as outer:
+        coro = work()
+    await coro
+
+    assert func_ctx is not None
+    assert func_ctx.get_parent_context(raise_if_none=False) is outer
+
+
+async def test_async_context_decorator_no_parent_when_called_outside_context() -> None:
+    """
+    Coroutine created outside any context, awaited inside one.
+    func_ctx should have no parent — the context active at await-time
+    is irrelevant.
+    """
+    func_ctx = None
+
+    @promising.context
+    async def work() -> str:
+        nonlocal func_ctx
+        func_ctx = promising.get_active_context()
+        return "done"
+
+    coro = work()
+    with promising.context():
+        await coro
+
+    assert func_ctx is not None
+    assert func_ctx.get_parent_context(raise_if_none=False) is None
+
+
+async def test_promise_has_parent_when_created_in_context() -> None:
+    """
+    A child Promise created inside a parent Promise's
+    execution has get_parent_context() and get_parent_promise() pointing to
+    the parent.
+    """
+    child_promise = None
+
+    @promising.function
+    async def child_func() -> str:
+        return "child"
+
+    @promising.function
+    async def parent_func() -> str:
+        nonlocal child_promise
+        child_promise = child_func()
+        return "parent"
+
+    parent_promise = parent_func()
+    await parent_promise
+
+    assert child_promise is not None
+    await child_promise
+    assert child_promise.get_parent_context(raise_if_none=False) is parent_promise
+    assert child_promise.get_parent_promise(raise_if_none=False) is parent_promise
+
+
+async def test_promise_has_no_parent_outside_context() -> None:
+    """
+    A Promise created at top level (outside any parent
+    context) has no parent.
+    """
+
+    @promising.function
+    async def noop() -> None:
+        pass
+
+    promise = noop()
+    assert promise.get_parent_context(raise_if_none=False) is None
+    assert promise.get_parent_promise(raise_if_none=False) is None
+    await promise
+
+
+async def test_active_promise_accessible_inside_sync_function() -> None:
+    """
+    get_active_promise() inside a sync promising function
+    (running in a thread pool) returns the wrapping Promise.
+    """
+
+    @promising.function(use_thread_pool=True)
+    def sync_func() -> promising.Promise:
+        return promising.get_active_promise(raise_if_none=False)
+
+    promise = sync_func()
+    current_from_inside = await promise.unpack_once()
+    assert current_from_inside is promise
+
+
+async def test_sync_parent_child_relationship() -> None:
+    """
+    A child Promise created inside a sync promising
+    function has the sync function's Promise as its parent.
+    """
+    child_promise = None
+
+    @promising.function
+    async def child_func() -> str:
+        return "child"
+
+    @promising.function(use_thread_pool=True)
+    def sync_parent() -> None:
+        nonlocal child_promise
+        child_promise = child_func(start_soon=False)
+
+    parent_promise = sync_parent()
+    await parent_promise
+
+    assert child_promise is not None
+    assert child_promise.get_parent_context(raise_if_none=False) is parent_promise
+    assert child_promise.get_parent_promise(raise_if_none=False) is parent_promise
+    await child_promise
