+++ tests/resolution/test_promising_function_run.py
@@ -102,50 +102,3 @@ def test_promising_function_run_with_child_promise(
         assert parent_work.run() == 42
 
     run_in_thread(_test)
-
-
-async def test_async_context_decorator_resolves_parent_at_call_site() -> None:
-    """
-    The parent of a @promising.context-decorated async function's context
-    is determined at call-site (when the coroutine object is created),
-    not at await-site (when the coroutine body runs).
-
-    Scenario: coroutine created inside `outer`, awaited outside it.
-    func_ctx should still have `outer` as its parent.
-    """
-    func_ctx = None
-
-    @promising.context
-    async def work() -> str:
-        nonlocal func_ctx
-        func_ctx = promising.get_active_context()
-        return "done"
-
-    with promising.context() as outer:
-        coro = work()
-    await coro
-
-    assert func_ctx is not None
-    assert func_ctx.get_parent_context(raise_if_none=False) is outer
-
-
-async def test_async_context_decorator_no_parent_when_called_outside_context() -> None:
-    """
-    Coroutine created outside any context, awaited inside one.
-    func_ctx should have no parent — the context active at await-time
-    is irrelevant.
-    """
-    func_ctx = None
-
-    @promising.context
-    async def work() -> str:
-        nonlocal func_ctx
-        func_ctx = promising.get_active_context()
-        return "done"
-
-    coro = work()
-    with promising.context():
-        await coro
-
-    assert func_ctx is not None
-    assert func_ctx.get_parent_context(raise_if_none=False) is None
