--- tests_old/promising_function/test_promising_function.py
@@ -1,44 +1,0 @@
-import promising
-
-
-async def test_promise_has_parent_when_created_in_context() -> None:
-    """
-    A child Promise created inside a parent Promise's
-    execution has get_parent_context() and get_parent_promise() pointing to
-    the parent.
-    """
-    child_promise = None
-
-    @promising.function
-    async def child_func() -> str:
-        return "child"
-
-    @promising.function
-    async def parent_func() -> str:
-        nonlocal child_promise
-        child_promise = child_func()
-        return "parent"
-
-    parent_promise = parent_func()
-    await parent_promise
-
-    assert child_promise is not None
-    await child_promise
-    assert child_promise.get_parent_context(raise_if_none=False) is parent_promise
-    assert child_promise.get_parent_promise(raise_if_none=False) is parent_promise
-
-
-async def test_promise_has_no_parent_outside_context() -> None:
-    """
-    A Promise created at top level (outside any parent
-    context) has no parent.
-    """
-
-    @promising.function
-    async def noop() -> None:
-        pass
-
-    promise = noop()
-    assert promise.get_parent_context(raise_if_none=False) is None
-    assert promise.get_parent_promise(raise_if_none=False) is None
-    await promise
